feat(api): move CRUDish modules to crud directory
