+++ packages/api/src/app.module.ts
@@ -4,16 +4,16 @@ import { EventEmitterModule } from '@nestjs/event-emitter';
 import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
 
-import { env } from '@/common/env';
+import { env } from './shared/env';
 
-import { AuthModule } from './auth/auth.module';
-import { CoursesModule } from './courses/courses.module';
+import { AuthModule } from './crud/auth/auth.module';
+import { CoursesModule } from './crud/courses/courses.module';
 import { SendEmailWhenLearningMaterialsUrlWasGeneratedAutomationModule } from './module/automation/send-email-when-learning-materials-url-was-generated/send-email-when-learning-materials-url-was-generated-automation.module';
 import { CourseProgressReadModule } from './module/read/course-progress/course-progress.read-module';
 import { LearningMaterialsReadModule } from './module/read/learning-materials/learning-materials.read-module';
 import { LearningMaterialsUrlWriteModule } from './module/write/learning-materials-url/learning-materials-url.write-module';
-import { PrismaModule } from './prisma/prisma.module';
-import { UsersModule } from './users/users.module';
+import { PrismaModule } from './shared/prisma/prisma.module';
+import { UsersModule } from './crud/users/users.module';
 
 const isProduction = env.NODE_ENV === 'production';
