+++ packages/api/src/crud/auth/auth.controller.ts
@@ -16,9 +16,9 @@ import type { LogoutResponse } from '@coderscamp/shared/models/auth';
 import { LoginResponse } from '@coderscamp/shared/models/auth/login';
 import { RegisterBody, registerError, RegisterResponse } from '@coderscamp/shared/models/auth/register';
 
-import { env } from '@/common/env';
+import { env } from '../../shared/env';
 
-import { isUniqueConstraintError } from '../prisma/prisma.errors';
+import { isUniqueConstraintError } from '../../shared/prisma/prisma.errors';
 import { fromUserToJwt } from './jwt/jwt.utils';
 import { LocalGuardRequest } from './local/local.types';
 import { LocalAuthGuard } from './local/local-auth.guard';
