+++ packages/api/src/crud/auth/events/user-registration-started.event.ts
