+++ packages/api/src/crud/auth/user-registration.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../../shared/prisma/prisma.service';
 
 @Injectable()
 export class UserRegistrationRepository {
