+++ packages/api/src/module/read/learning-materials/learning-materials.read-module.spec.ts
@@ -3,9 +3,9 @@ import { AsyncReturnType } from 'type-fest';
 import { v4 as uuid } from 'uuid';
 import waitForExpect from 'wait-for-expect';
 
-import { initReadTestModule } from '@/common/test-utils';
+import { initReadTestModule } from '../../../shared/test-utils';
 import { LearningMaterialsUrlWasGenerated } from '@/events/learning-materials-url-was-generated.domain-event';
-import { UserId } from '@/users/users.types';
+import { UserId } from '@/crud/users/users.types';
 import { StorableEvent } from '@/write/shared/application/event-repository';
 import { EventStreamName } from '@/write/shared/application/event-stream-name.value-object';
