+++ packages/api/src/module/read/learning-materials/learning-materials.read-module.ts
@@ -3,7 +3,7 @@ import { OnEvent } from '@nestjs/event-emitter';
 
 import { LearningMaterialsUrlWasGenerated } from '@/events/learning-materials-url-was-generated.domain-event';
 import { ApplicationEvent } from '@/module/application-command-events';
-import { PrismaService } from '@/prisma/prisma.service';
+import { PrismaService } from '../../../shared/prisma/prisma.service';
 import { SharedModule } from '@/write/shared/shared.module';
 
 import { LearningMaterialsRestController } from './learning-materials.rest-controller';
