+++ packages/api/src/module/write/learning-materials-url/application/users.port.ts
@@ -1,4 +1,4 @@
-import type { UserId } from '@/users/users.types';
+import type { UserId } from '@/crud/users/users.types';
 
 export const USERS_PORT = Symbol('USERS_PORT');
