+++ packages/api/src/module/write/learning-materials-url/learning-materials-url.write-module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
-import { UsersModule } from '@/users/users.module';
-import { UsersService } from '@/users/users.service';
+import { UsersModule } from '@/crud/users/users.module';
+import { UsersService } from '@/crud/users/users.service';
 import { SharedModule } from '@/write/shared/shared.module';
 
 import { GenerateLearningMaterialsUrlCommandHandler } from './application/generate-learning-materials-url.command-handler';
