+++ packages/api/src/module/write/shared/infrastructure/event-repository/prisma-event-repository.spec.ts
@@ -2,9 +2,9 @@ import { Test } from '@nestjs/testing';
 import { AsyncReturnType } from 'type-fest';
 import { v4 as uuid } from 'uuid';
 
-import { cleanupDatabase, sequence, storableEvent } from '@/common/test-utils';
-import { PrismaModule } from '@/prisma/prisma.module';
-import { PrismaService } from '@/prisma/prisma.service';
+import { cleanupDatabase, sequence, storableEvent } from '../../../../../shared/test-utils';
+import { PrismaModule } from '../../../../../shared/prisma/prisma.module';
+import { PrismaService } from '../../../../../shared/prisma/prisma.service';
 import { StorableEvent } from '@/write/shared/application/event-repository';
 import { EventStreamName } from '@/write/shared/application/event-stream-name.value-object';
 import { PrismaEventRepository } from '@/write/shared/infrastructure/event-repository/prisma-event-repository.service';
